requirements.txt
# Dev environment
pip
autopep8

# App
g4f
python-telegram-bot
flask